groq_transcribe/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
soundfile==0.12.1
mutagen==1.47.0
librosa==0.9.2
pydub==0.25.1
click==8.1.7
numpy>=1.19.5
scipy>=1.5.4
